分离接收消息、文件处理相关逻辑到MessageListener 中，初步支持同一个 webSocket 多次复用
